Standardizing types both inside dbscan and in the sparse csr prims

+++ cpp/src/datasets/make_blobs.cu
@@ -23,8 +23,8 @@
 namespace ML {
 namespace Datasets {
 
-void make_blobs(const cumlHandle& handle, float* out, int* labels, int n_rows,
-                int n_cols, int n_clusters, const float* centers,
+void make_blobs(const cumlHandle& handle, float* out, long* labels, long n_rows,
+                long n_cols, long n_clusters, const float* centers,
                 const float* cluster_std, const float cluster_std_scalar,
                 bool shuffle, float center_box_min, float center_box_max,
                 uint64_t seed) {
@@ -34,11 +34,11 @@ void make_blobs(const cumlHandle& handle, float* out, int* labels, int n_rows,
                                shuffle, center_box_min, center_box_max, seed);
 }
 
-void make_blobs(const cumlHandle& handle, double* out, int* labels, int n_rows,
-                int n_cols, int n_clusters, const double* centers,
-                const double* cluster_std, const double cluster_std_scalar,
-                bool shuffle, double center_box_min, double center_box_max,
-                uint64_t seed) {
+void make_blobs(const cumlHandle& handle, double* out, long* labels,
+                long n_rows, long n_cols, long n_clusters,
+                const double* centers, const double* cluster_std,
+                const double cluster_std_scalar, bool shuffle,
+                double center_box_min, double center_box_max, uint64_t seed) {
   MLCommon::Random::make_blobs(out, labels, n_rows, n_cols, n_clusters,
                                handle.getDeviceAllocator(), handle.getStream(),
                                centers, cluster_std, cluster_std_scalar,
