+++ cpp/src/dbscan/dbscan_api.h
@@ -40,12 +40,11 @@ extern "C" {
  * @{
  */
 cumlError_t cumlSpDbscanFit(cumlHandle_t handle, float *input, long n_rows,
-                            long n_cols, float eps, int min_pts, int *labels,
+                            long n_cols, float eps, int min_pts, long *labels,
                             size_t max_bytes_per_batch, int verbose);
 cumlError_t cumlDpDbscanFit(cumlHandle_t handle, double *input, long n_rows,
-                            long n_cols, double eps, int min_pts,
-                            int *labels, size_t max_bytes_per_batch,
-                            int verbose);
+                            long n_cols, double eps, int min_pts, long *labels,
+                            size_t max_bytes_per_batch, int verbose);
 /** @} */
 
 #ifdef __cplusplus
